Resolve conflicts and complete cherry-pick of commit f6553a8
- Merged changes from f6553a8, including updates to backend services and configuration files.
- Added new diagram files and documentation for better project understanding.
- Resolved conflicts in multiple files, ensuring compatibility with the current branch.

+++ backend/middleware/db-check.js
@@ -103,7 +103,7 @@ async function createProposalsTable() {
             console.log("Creating proposals table...")
             await pool.query(`
         CREATE TABLE proposals (
-          id INT AUTO_INCREMENT PRIMARY KEY,
+          id BIGINT AUTO_INCREMENT PRIMARY KEY,
           title VARCHAR(255) NOT NULL,
           description TEXT NOT NULL,
           category VARCHAR(100) NOT NULL,
@@ -177,7 +177,7 @@ async function createReviewsTable() {
             await pool.query(`
         CREATE TABLE reviews (
           id INT AUTO_INCREMENT PRIMARY KEY,
-          proposal_id INT NOT NULL,
+          proposal_id BIGINT NOT NULL,
           reviewer_id INT NOT NULL,
           comments TEXT NOT NULL,
           rating INT NOT NULL,
